requirements/base.txt
dj-database-url==0.5.0
Django==2.1.10
psycopg2-binary==2.8.3
django-admin==1.3.2
djangorestframework==3.9.4
djangorestframework-camel-case
django-cors-headers
django-rest-knox
